Fix IT suite.
